[jsdom] Use `base.d.ts` file
