+++ types/jsdom/index.d.ts
@@ -8,7 +8,7 @@
 /// <reference path="ts3.5/index.d.ts"/>
 
 // tslint:disable-next-line: no-declare-current-package no-single-declare-module
-declare module 'jsdom' {
+declare module "jsdom" {
     interface DOMWindow {
         InputEvent: typeof InputEvent;
         External: typeof External;
