+++ types/jsdom/jsdom-tests.ts
@@ -1,10 +1,10 @@
-import './test/core';
-import jsdom = require('jsdom');
+import "./test/core";
+import jsdom = require("jsdom");
 
 const dom = new jsdom.JSDOM();
 const domWindow = dom.window; // $ExpectType DOMWindow
 
-domWindow.document.querySelector('slot'); // $ExpectType HTMLSlotElement | null
+domWindow.document.querySelector("slot"); // $ExpectType HTMLSlotElement | null
 domWindow.AbstractRange.prototype; // $ExpectType AbstractRange
 domWindow.StaticRange.prototype; // $ExpectType StaticRange
 domWindow.ShadowRoot.prototype; // $ExpectType ShadowRoot
@@ -15,4 +15,6 @@ domWindow.BigInt64Array; // $ExpectType BigInt64ArrayConstructor
 domWindow.BigUint64Array; // $ExpectType BigUint64ArrayConstructor
 domWindow.SharedArrayBuffer; // $ExpectType SharedArrayBufferConstructor
 domWindow.WebAssembly; // $ExpectType typeof WebAssembly
+
+domWindow.InputEvent.prototype; // $ExpectType InputEvent
 domWindow.external; // $ExpectType External
