+++ types/jsdom/package.json
@@ -3,9 +3,7 @@
     "types": "index",
     "typesVersions": {
         "<=3.5": {
-            "*": [
-                "ts3.5/*"
-            ]
+            "*": ["ts3.5/*"]
         }
     }
 }
