+++ types/jsdom/test/core.ts
@@ -8,13 +8,13 @@ import {
     ResourceLoader,
     FetchOptions,
     ConstructorOptions,
-} from 'jsdom';
-import { CookieJar as ToughCookieJar, MemoryCookieStore } from 'tough-cookie';
-import { Script } from 'vm';
+} from "jsdom";
+import { CookieJar as ToughCookieJar, MemoryCookieStore } from "tough-cookie";
+import { Script } from "vm";
 
 function test_basic_usage() {
     const dom = new JSDOM(`<!DOCTYPE html><p>Hello world</p>`);
-    console.log(dom.window.document.querySelector('p')!.textContent); // "Hello world"
+    console.log(dom.window.document.querySelector("p")!.textContent); // "Hello world"
 
     const { window } = new JSDOM(`...`);
     // or even
@@ -35,7 +35,7 @@ function test_executing_scripts2() {
         `<body>
   <script>document.body.appendChild(document.createElement("hr"));</script>
 </body>`,
-        { runScripts: 'dangerously' },
+        { runScripts: "dangerously" },
     );
 
     // The script will be executed and modify the DOM:
@@ -43,7 +43,7 @@ function test_executing_scripts2() {
 }
 
 function test_executing_scripts3() {
-    const window = new JSDOM(``, { runScripts: 'outside-only' }).window;
+    const window = new JSDOM(``, { runScripts: "outside-only" }).window;
 
     window.eval(`document.body.innerHTML = "<p>Hello, world!</p>";`);
     window.document.body.children.length === 1;
@@ -53,10 +53,10 @@ function test_virtualConsole() {
     const virtualConsole = new VirtualConsole();
     const dom = new JSDOM(``, { virtualConsole });
 
-    virtualConsole.on('error', () => {});
-    virtualConsole.on('warn', () => {});
-    virtualConsole.on('info', () => {});
-    virtualConsole.on('dir', () => {});
+    virtualConsole.on("error", () => {});
+    virtualConsole.on("warn", () => {});
+    virtualConsole.on("info", () => {});
+    virtualConsole.on("dir", () => {});
     // ... etc. See https://console.spec.whatwg.org/#logging
 
     virtualConsole.sendTo(console);
@@ -80,21 +80,21 @@ function test_beforeParse() {
 }
 
 function test_storageQuota() {
-    new JSDOM('', { storageQuota: 1337 });
+    new JSDOM("", { storageQuota: 1337 });
 }
 
 function test_pretendToBeVisual() {
-    new JSDOM('', { pretendToBeVisual: true });
+    new JSDOM("", { pretendToBeVisual: true });
 }
 
 function test_serialize() {
     const dom = new JSDOM(`<!DOCTYPE html>hello`);
 
-    dom.serialize() === '<!DOCTYPE html><html><head></head><body>hello</body></html>';
+    dom.serialize() === "<!DOCTYPE html><html><head></head><body>hello</body></html>";
 
     // Contrast with:
     // tslint:disable-next-line no-unnecessary-type-assertion
-    dom.window.document.documentElement!.outerHTML === '<html><head></head><body>hello</body></html>';
+    dom.window.document.documentElement!.outerHTML === "<html><head></head><body>hello</body></html>";
 }
 
 function test_nodeLocation() {
@@ -107,9 +107,9 @@ function test_nodeLocation() {
 
     const document = dom.window.document;
     const bodyEl = document.body; // implicitly created
-    const pEl = document.querySelector('p')!;
+    const pEl = document.querySelector("p")!;
     const textNode = pEl.firstChild!;
-    const imgEl = document.querySelector('img')!;
+    const imgEl = document.querySelector("img")!;
 
     console.log(dom.nodeLocation(bodyEl)); // null; it's not in the source
     console.log(dom.nodeLocation(pEl)); // { startOffset: 0, endOffset: 39, startTag: ..., endTag: ... }
@@ -118,7 +118,7 @@ function test_nodeLocation() {
 }
 
 function test_runVMScript() {
-    const dom = new JSDOM(``, { runScripts: 'outside-only' });
+    const dom = new JSDOM(``, { runScripts: "outside-only" });
     const script = new Script(`
     if (!this.ran) {
         this.ran = 0;
@@ -139,30 +139,30 @@ function test_reconfigure(myFakeTopForTesting: DOMWindow) {
     const dom = new JSDOM();
 
     dom.window.top === dom.window;
-    dom.window.location.href === 'about:blank';
+    dom.window.location.href === "about:blank";
 
-    dom.reconfigure({ windowTop: myFakeTopForTesting, url: 'https://example.com/' });
+    dom.reconfigure({ windowTop: myFakeTopForTesting, url: "https://example.com/" });
 
     dom.window.top === myFakeTopForTesting;
-    dom.window.location.href === 'https://example.com/';
+    dom.window.location.href === "https://example.com/";
 }
 
 function test_fromURL() {
     const options: BaseOptions = {};
 
-    JSDOM.fromURL('https://example.com/', options).then(dom => {
+    JSDOM.fromURL("https://example.com/", options).then(dom => {
         console.log(dom.serialize());
     });
 
     function pretendToBeVisual() {
-        JSDOM.fromURL('https://github.com', {
+        JSDOM.fromURL("https://github.com", {
             pretendToBeVisual: true,
         });
     }
 }
 
 function test_fromFile(options: FileOptions) {
-    JSDOM.fromFile('stuff.html', options).then(dom => {
+    JSDOM.fromFile("stuff.html", options).then(dom => {
         console.log(dom.serialize());
     });
 }
@@ -171,7 +171,7 @@ function test_fragment() {
     const frag = JSDOM.fragment(`<p>Hello</p><p><strong>Hi!</strong>`);
 
     frag.childNodes.length === 2;
-    frag.querySelector('strong')!.textContent = 'Why hello there!';
+    frag.querySelector("strong")!.textContent = "Why hello there!";
     // etc.
 }
 
@@ -199,5 +199,5 @@ function test_custom_resource_loader() {
             return super.fetch(url, options);
         }
     }
-    new JSDOM('', { resources: new CustomResourceLoader() });
+    new JSDOM("", { resources: new CustomResourceLoader() });
 }
