+++ types/jsdom/ts3.5/jsdom-tests.ts
@@ -1,10 +1,10 @@
-import '../test/core';
-import jsdom = require('jsdom');
+import "../test/core";
+import jsdom = require("jsdom");
 
 const dom = new jsdom.JSDOM();
 const domWindow = dom.window; // $ExpectType DOMWindow
 
-domWindow.document.querySelector('slot'); // $ExpectType HTMLSlotElement | null
+domWindow.document.querySelector("slot"); // $ExpectType HTMLSlotElement | null
 domWindow.AbstractRange.prototype; // $ExpectType AbstractRange
 domWindow.StaticRange.prototype; // $ExpectType StaticRange
 domWindow.ShadowRoot.prototype; // $ExpectType ShadowRoot
