+++ types/jsdom/ts3.5/tsconfig.json
@@ -1,7 +1,7 @@
 {
     "compilerOptions": {
         "module": "commonjs",
-        "lib": ["es2020", "dom"],
+        "lib": ["ESNext", "DOM"],
         "noImplicitAny": true,
         "noImplicitThis": true,
         "strictNullChecks": true,
